All functions moved to the neo anim library
